Move XAS editors from owpreprocess.py
