+++ orangecontrib/spectroscopy/widgets/preprocessors/__init__.py
@@ -7,3 +7,4 @@ from . import emsc
 from . import me_emsc
 from . import normalize
 from . import spikeremoval
+from . import xas
